Fix bad refactor for race type file
